[WIP] Remaking joystick display example
Signed off by: @josealvim
Signed off by: @icemage144
